requirements.txt
asgiref==3.11.0
bleach==6.3.0
boto3==1.42.50
botocore==1.42.50
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.1.1
dj-rest-auth==5.0.2
Django==4.2.17
django-allauth==0.59.0
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-countries==7.6.1
django-js-asset==3.1.2
django-storages==1.14.6
django-summernote==0.8.20.0
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
jmespath==1.1.0
oauthlib==3.3.1
packaging==25.0
pillow==10.2.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==2.0.0
redis==5.2.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
stripe==14.0.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.11.0
